requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.3
